Move unique to external.utils.
