+++ components/server/src/external/model/transformations.py
@@ -4,10 +4,9 @@ import json
 
 from collections.abc import Iterator
 from json.decoder import JSONDecodeError
-from server_utilities.functions import unique
 from server_utilities.type import EditScope, ItemId
 
-from ..utils.functions import asymmetric_decrypt, asymmetric_encrypt, uuid, DecryptionError
+from ..utils.functions import asymmetric_decrypt, asymmetric_encrypt, unique, uuid, DecryptionError
 
 from .iterators import sources as iter_sources
 from .queries import is_password_parameter
